% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_google_profile.R
\name{scrape_google_profile}
\alias{scrape_google_profile}
\title{Web Scraping for 'Google Scholar' Profiles}
\usage{
scrape_google_profile(url)
}
\arguments{
\item{url}{A url link of the google scholar profile.}
}
\value{
A dataframe containing publishing information such as title, author, journal, date, page, and so on.
}
\description{
Get all of the published information from the profile page by using the 'Google Scholar' profiles link and putting it into a table.
}
\examples{
url <- "https://scholar.google.com/citations?user=hGTnO4oAAAAJ&hl=en&authuser=3"
paper_list <- scrape_google_profile(url)
}
