% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_pub.R
\name{scrape_pub}
\alias{scrape_pub}
\title{Web Scraping for 'PubMed'}
\usage{
scrape_pub(first_name, last_name)
}
\arguments{
\item{first_name}{The author's first name}

\item{last_name}{The author's last name}
}
\value{
A dataframe containing publishing information such as title, author, journal, date, page, and so on.
}
\description{
Use web scraping for 'PubMed' searches based on first and last names to get all of the published information from the search page. Then placing it on a table
}
\examples{
paper_list <- scrape_pub("Jedi","wars")
}
